refactorSOIndex
